refactor(builders): share select generators across all three dialects
Deletes pg's plain-select fallback and unifies the read side.

The fallback is unreachable under drizzle-orm v1. Two independent proofs:

  - `resolveQueryExecutor` can only return a falsy `queryBase` when the
    build-time one was already falsy, and a falsy build-time `queryBase`
    now throws `missingQueryBuilderError` instead of silently degrading.
  - `PgAsyncDatabase` no longer carries `_.fullSchema` (removed in rc.2),
    so `buildSchema` derives PG's schema from `_.relations`. PG's schema
    and `db.query` therefore cannot disagree: every table in the generated
    schema has a query builder by construction.

With the fallback gone, pg's two select generators are identical to
sqlite's. MySQL's turned out identical as well — the only differences were
a type annotation, one property's position in an object literal, and a
comment. All three now come from `createSelectGenerators`, parameterized
by the two things that actually vary: the dialect's primary-key extractor
and its null ordering.

That also removes `generateSelectArray`/`generateSelectSingle` from
`DialectSchemaAdapter`: both are derivable from `primaryKeyPropNames` +
`nullOrdering`, which the adapter already supplies.

  pg.ts      350 -> 31
  sqlite.ts  289 -> 130
  mysql.ts  1365 -> 1231
  select.ts    0 -> 208
