+++ src/util/builders/pg.ts
@@ -1,327 +1,10 @@
-import { and, getColumns, type Table } from 'drizzle-orm';
-import type { RelationalQueryBuilder } from 'drizzle-orm/mysql-core/query-builders/query';
-import { getTableConfig, type PgAsyncDatabase, type PgColumn, PgTable } from 'drizzle-orm/pg-core';
-import { GraphQLError, type GraphQLInputObjectType } from 'graphql';
-import type { ResolveTree } from 'graphql-parse-resolve-info';
-import { parseResolveInfo } from 'graphql-parse-resolve-info';
+import type { Table } from 'drizzle-orm';
+import { getTableConfig, type PgAsyncDatabase, PgTable } from 'drizzle-orm/pg-core';
 import type { GeneratedEntities } from '../../types.ts';
-import {
-  applyLimitPolicy,
-  attachRowCursors,
-  buildCursorCondition,
-  type CursorOrderEntry,
-  cursorOrderExprs,
-  cursorOrderingEntries,
-  type DeletedMode,
-  decodeCursor,
-  extractFilters,
-  extractOrderBy,
-  extractSelectedColumnsFromTreeSQLFormat,
-  generateDistinctEnum,
-  getPrimaryKeyPropNamesFromConfig,
-  isCursorFieldSelected,
-  type LimitPolicyFor,
-  orderByCursorObstacle,
-  primaryKeyOrderExprs,
-  primaryKeyRestriction,
-  type RelationFilterBase,
-  type ResolverPolicies,
-  relationFilterCtx,
-  resolveQueryExecutor,
-  runRelationalSelect,
-  selectArrayArgs,
-  selectDistinctKeys,
-  selectSingleArgs,
-  type TablesRelationalConfig,
-  type TypeNameMapper,
-  toGraphQLError,
-  withDefaultOrderBy,
-  withScope,
-} from '../builders/common.ts';
-import { remapToGraphQLArrayOutput, remapToGraphQLSingleOutput } from '../data-mappers/index.ts';
+import { getPrimaryKeyPropNamesFromConfig, type TablesRelationalConfig } from '../builders/common.ts';
 import { createSchemaDataGenerator } from './schema-data.ts';
-import type { CreatedResolver, SchemaGeneratorOptions, TableNamedRelations, TableSelectArgs } from './types.ts';
+import type { SchemaGeneratorOptions } from './types.ts';
 import { createUpdateManyGenerator } from './update-many.ts';
-
-const generateSelectArray = (
-  db: PgAsyncDatabase<any, any>,
-  tableName: string,
-  tables: Record<string, Table>,
-  relationMap: Record<string, Record<string, TableNamedRelations>>,
-  orderArgs: GraphQLInputObjectType,
-  filterArgs: GraphQLInputObjectType,
-  fieldName: string,
-  typeName: string,
-  typeNameMapper?: TypeNameMapper,
-  filterCtx?: RelationFilterBase,
-  distinctEnabled: boolean = true,
-  limits?: LimitPolicyFor,
-  policies?: ResolverPolicies,
-): CreatedResolver => {
-  const queryBase = db.query[tableName as keyof typeof db.query & string] as unknown as
-    | RelationalQueryBuilder<any, any, any>
-    | undefined;
-  // Tables without relations won't have db.query support — fall back to basic select.
-
-  const table = tables[tableName]!;
-  const limitPolicy = limits?.(tableName);
-  const pkNames = pgPrimaryKeyPropNames(table as PgTable);
-  const queryArgs = selectArrayArgs(
-    orderArgs,
-    filterArgs,
-    distinctEnabled ? generateDistinctEnum(table, typeName) : undefined,
-    policies?.softDelete,
-    tableName,
-  );
-
-  return {
-    name: fieldName,
-    resolver: async (_source, rawArgs: Partial<TableSelectArgs>, context, info) => {
-      // An omitted `orderBy` falls back to the table's configured default ordering here,
-      // before anything reads the arguments — the cursor tuple, a `distinct` pass and the
-      // plain-select fallback all have to agree on one effective ordering.
-      const args = withDefaultOrderBy(rawArgs, tableName, policies?.defaultOrderBy);
-      try {
-        const parsedInfo = parseResolveInfo(info, { deep: true }) as ResolveTree;
-        const { executor, queryBase: requestQueryBase } = resolveQueryExecutor(db, context, tableName, queryBase);
-        const scope = policies?.scope?.(context);
-        // Resolved once here so the relational path and the plain-select fallback below both
-        // see the same effective limit.
-        const limit = applyLimitPolicy(args.limit, limitPolicy, fieldName);
-
-        if (requestQueryBase) {
-          return await runRelationalSelect({
-            queryBase: requestQueryBase,
-            tables,
-            tableName,
-            table,
-            relationMap,
-            typeName,
-            typeNameMapper,
-            parsedInfo,
-            ...args,
-            limit,
-            single: false,
-            filterCtx,
-            limits,
-            defaultOrderBy: policies?.defaultOrderBy,
-            pkNames,
-            db: executor,
-            scope,
-            // PostgreSQL sorts NULLs as the largest values (last in ASC).
-            nullOrdering: 'nulls-largest',
-          });
-        }
-
-        // Fallback for tables without relational query builder support.
-        // Use SQL column objects (not Record<string,true>) so db.select() receives valid expressions.
-        const { offset, orderBy, where, distinct, after, deleted } = args as Partial<TableSelectArgs> & {
-          deleted?: DeletedMode;
-        };
-        const selectedColumnsSql = extractSelectedColumnsFromTreeSQLFormat<PgColumn>(
-          parsedInfo.fieldsByTypeName[typeName]!,
-          table,
-          { tableName, relationMap, tables },
-        );
-
-        // Keyset pagination (see runRelationalSelect, which handles the RQB path).
-        const cursorSelected = isCursorFieldSelected(parsedInfo.fieldsByTypeName[typeName], table);
-        let cursorEntries: CursorOrderEntry[] | undefined;
-        if (after != null || cursorSelected) {
-          if (after != null && distinct?.length) {
-            throw new GraphQLError("'after' cannot be combined with 'distinct'.");
-          }
-          const cursorObstacle = orderByCursorObstacle(orderBy);
-          if (cursorObstacle) {
-            if (after != null) {
-              throw new GraphQLError(cursorObstacle);
-            }
-            // `cursor` selected under an ordering a cursor cannot express — resolves to null;
-            // the ordering itself still applies.
-          } else if (!pkNames.length) {
-            if (after != null) {
-              throw new GraphQLError(
-                `Table ${tableName} has no primary key, so cursor pagination cannot be used on it.`,
-              );
-            }
-          } else {
-            cursorEntries = cursorOrderingEntries(orderBy, pkNames);
-            if (cursorSelected) {
-              // The whole ordering tuple is needed to compute each row's cursor.
-              const allColumns = getColumns(table);
-              for (const [column] of cursorEntries) {
-                selectedColumnsSql[column] ??= allColumns[column] as PgColumn;
-              }
-            }
-          }
-        }
-        const cursorCondition =
-          after != null && cursorEntries
-            ? buildCursorCondition(table, cursorEntries, decodeCursor(after, cursorEntries), 'nulls-largest')
-            : undefined;
-
-        const baseWhereSql = withScope(
-          scope,
-          tableName,
-          table,
-          where ? extractFilters(table, tableName, where, relationFilterCtx(filterCtx, tableName)) : undefined,
-          deleted,
-        );
-        const whereSql = cursorCondition ? and(baseWhereSql, cursorCondition) : baseWhereSql;
-
-        // `distinct` picks the surviving rows in its own pass; the main query is then narrowed
-        // to those primary keys and re-orders them the same way. See runRelationalSelect.
-        let distinctKeys: Record<string, any>[] | undefined;
-        if (distinct?.length) {
-          distinctKeys = await selectDistinctKeys({
-            db: executor,
-            table,
-            tableName,
-            distinct,
-            pkNames,
-            where: whereSql,
-            orderBy,
-            limit,
-            offset,
-          });
-          if (!distinctKeys.length) {
-            return [];
-          }
-        }
-
-        let q = executor.select(selectedColumnsSql).from(table);
-        if (distinctKeys) {
-          q = q.where(primaryKeyRestriction(table, pkNames, distinctKeys)) as any;
-        } else if (whereSql) {
-          q = q.where(whereSql) as any;
-        }
-        if (cursorEntries && !distinctKeys) {
-          // Cursor pagination pages over a total order: orderBy plus the PK tiebreak.
-          q = q.orderBy(...cursorOrderExprs(table, cursorEntries)) as any;
-        } else if (orderBy) {
-          q = q.orderBy(
-            ...extractOrderBy(table, orderBy, relationFilterCtx(filterCtx, tableName), where),
-            ...(distinctKeys ? primaryKeyOrderExprs(table, pkNames) : []),
-          ) as any;
-        } else if ((distinctKeys || offset != null || limit != null) && pkNames.length) {
-          // See runRelationalSelect: an unordered slice is not stable between requests.
-          q = q.orderBy(...primaryKeyOrderExprs(table, pkNames)) as any;
-        }
-        if (!distinctKeys) {
-          if (offset) {
-            q = q.offset(offset) as any;
-          }
-          if (limit) {
-            q = q.limit(limit) as any;
-          }
-        }
-        const rows = await q;
-        if (cursorEntries && cursorSelected) {
-          attachRowCursors(rows, cursorEntries);
-        }
-        return remapToGraphQLArrayOutput(rows, tableName, table, relationMap);
-      } catch (e) {
-        throw toGraphQLError(e);
-      }
-    },
-    args: queryArgs,
-  };
-};
-
-const generateSelectSingle = (
-  db: PgAsyncDatabase<any, any>,
-  tableName: string,
-  tables: Record<string, Table>,
-  relationMap: Record<string, Record<string, TableNamedRelations>>,
-  orderArgs: GraphQLInputObjectType,
-  filterArgs: GraphQLInputObjectType,
-  fieldName: string,
-  typeName: string,
-  typeNameMapper?: TypeNameMapper,
-  filterCtx?: RelationFilterBase,
-  limits?: LimitPolicyFor,
-  policies?: ResolverPolicies,
-): CreatedResolver => {
-  const queryBase = db.query[tableName as keyof typeof db.query & string] as unknown as
-    | RelationalQueryBuilder<any, any, any>
-    | undefined;
-  // Tables without relations won't have db.query support — fall back to basic select.
-
-  const queryArgs = selectSingleArgs(orderArgs, filterArgs, policies?.softDelete, tableName);
-
-  const table = tables[tableName]!;
-  const pkNames = pgPrimaryKeyPropNames(table as PgTable);
-
-  return {
-    name: fieldName,
-    resolver: async (_source, rawArgs: Partial<TableSelectArgs>, context, info) => {
-      // An omitted `orderBy` falls back to the table's configured default ordering here,
-      // before anything reads the arguments — the cursor tuple, a `distinct` pass and the
-      // plain-select fallback all have to agree on one effective ordering.
-      const args = withDefaultOrderBy(rawArgs, tableName, policies?.defaultOrderBy);
-      try {
-        const parsedInfo = parseResolveInfo(info, { deep: true }) as ResolveTree;
-        const { executor, queryBase: requestQueryBase } = resolveQueryExecutor(db, context, tableName, queryBase);
-        const scope = policies?.scope?.(context);
-
-        if (requestQueryBase) {
-          return await runRelationalSelect({
-            queryBase: requestQueryBase,
-            tables,
-            tableName,
-            table,
-            relationMap,
-            typeName,
-            typeNameMapper,
-            parsedInfo,
-            ...args,
-            single: true,
-            filterCtx,
-            limits,
-            defaultOrderBy: policies?.defaultOrderBy,
-            pkNames,
-            db: executor,
-            scope,
-          });
-        }
-
-        // Fallback for tables without relational query builder support.
-        const { offset, orderBy, where, deleted } = args as Partial<TableSelectArgs> & { deleted?: DeletedMode };
-        const selectedColumnsSql = extractSelectedColumnsFromTreeSQLFormat<PgColumn>(
-          parsedInfo.fieldsByTypeName[typeName]!,
-          table,
-          { tableName, relationMap, tables },
-        );
-        let q = executor.select(selectedColumnsSql).from(table);
-        const whereSql = withScope(
-          scope,
-          tableName,
-          table,
-          where ? extractFilters(table, tableName, where, relationFilterCtx(filterCtx, tableName)) : undefined,
-          deleted,
-        );
-        if (whereSql) {
-          q = q.where(whereSql) as any;
-        }
-        if (orderBy) {
-          q = q.orderBy(...extractOrderBy(table, orderBy, relationFilterCtx(filterCtx, tableName), where)) as any;
-        } else if (pkNames.length) {
-          // A single query is an implicit `limit 1` — order it so the row is deterministic.
-          q = q.orderBy(...primaryKeyOrderExprs(table, pkNames)) as any;
-        }
-        if (offset) {
-          q = q.offset(offset) as any;
-        }
-        const rows = await q.limit(1);
-        const result = rows[0];
-        return result ? remapToGraphQLSingleOutput(result, tableName, table, relationMap) : undefined;
-      } catch (e) {
-        throw toGraphQLError(e);
-      }
-    },
-    args: queryArgs,
-  };
-};
 
 /** Primary-key property names for a PG table, including table-level composite keys. */
 const pgPrimaryKeyPropNames = (table: PgTable): string[] => getPrimaryKeyPropNamesFromConfig(table, getTableConfig);
@@ -334,8 +17,6 @@ const pgSchemaData = createSchemaDataGenerator({
   primaryKeyPropNames: pgPrimaryKeyPropNames,
   // PostgreSQL sorts NULLs as the largest values (last in ASC).
   nullOrdering: 'nulls-largest',
-  generateSelectArray,
-  generateSelectSingle,
   generateUpdateMany,
 });
